Add additional methods to toggle views and set date defaults to domUpdates object
